Remove dwsync.xml in web/_notes
It seems like this points toward something that attempts to
ensure the state of the wp-config.php file. We'll be generating
that file automatically during provisioning, so shouldn't need
anything special.

If CBN or one of the plugins needs an extra configuration in
wp-config, we'll need to figure out the best way to handle it.